Relocated determine_top_group_indices to common
from conformers
