+++ arc/commonTest.py
@@ -10,6 +10,8 @@ import os
 import time
 import unittest
 
+from rmgpy.molecule.molecule import Molecule
+
 import arc.common as common
 from arc.exceptions import InputError, SettingsError
 from arc.settings import arc_path, servers
@@ -104,6 +106,31 @@ class TestCommon(unittest.TestCase):
         with self.assertRaises(SettingsError):
             ess_settings7 = {'gaussian': ['noserver']}
             common.check_ess_settings(ess_settings7)
+
+    def test_determine_top_group_indices(self):
+        """Test determining the top group in a molecule"""
+        mol = Molecule(smiles='c1cc(OC)ccc1OC(CC)SF')
+        atom1 = mol.atoms[9]  # this is the C atom at the S, O, H, and C junction
+        atom2a = mol.atoms[10]  # C
+        atom2b = mol.atoms[8]  # O
+        atom2c = mol.atoms[12]  # S
+        atom2d = mol.atoms[21]  # H
+
+        top, top_has_heavy_atoms = common.determine_top_group_indices(mol, atom1, atom2a)
+        self.assertEqual(len(top), 7)
+        self.assertTrue(top_has_heavy_atoms)
+
+        top, top_has_heavy_atoms = common.determine_top_group_indices(mol, atom1, atom2b)
+        self.assertEqual(len(top), 16)
+        self.assertTrue(top_has_heavy_atoms)
+
+        top, top_has_heavy_atoms = common.determine_top_group_indices(mol, atom1, atom2c)
+        self.assertEqual(len(top), 2)
+        self.assertTrue(top_has_heavy_atoms)
+
+        top, top_has_heavy_atoms = common.determine_top_group_indices(mol, atom1, atom2d)
+        self.assertEqual(top, [22])
+        self.assertFalse(top_has_heavy_atoms)  # H
 
     def test_min_list(self):
         """Test the min_list() function"""
